Move initial and parameter
