+++ mira/sources/bilayer.py
@@ -3,7 +3,7 @@ of models based on mass-action kinetics."""
 import json
 import sympy
 
-from mira.metamodel.template_model import TemplateModel
+from mira.metamodel.template_model import TemplateModel, Parameter
 from mira.metamodel.templates import *
