+++ mira/sources/sbml.py
@@ -25,11 +25,9 @@ from mira.metamodel import (
     NaturalConversion,
     NaturalDegradation,
     NaturalProduction,
-    Parameter,
     Template,
 )
-from mira.metamodel.templates import Initial
-from mira.metamodel.template_model import TemplateModel
+from mira.metamodel.template_model import TemplateModel, Initial, Parameter
 
 __all__ = [
     "template_model_from_sbml_file",
