+++ tests/test_modeling/test_ode.py
@@ -5,8 +5,8 @@ import unittest
 import numpy
 
 from mira.metamodel import NaturalConversion, ControlledConversion, Concept
-from mira.metamodel.template_model import TemplateModel
-from mira.modeling import Model, Initial
+from mira.metamodel.template_model import TemplateModel, Initial
+from mira.modeling import Model
 from mira.modeling.ode import OdeModel, simulate_ode_model
